backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-openai
langchain-chroma
langchain-community
langchain-experimental
unstructured[all-docs]
htmltabletomd
openai
redis
chroma
Pillow
pytesseract
pdfplumber
speechrecognition
pydub
beautifulsoup4
PyPDF2
nltk
transformers
torch
spacy>=3.0.0
pdfminer.six
python-docx
docx2txt
matplotlib
PyMuPDF
camelot-py[cv]
